backend/requirements.txt
# AI Academic Timetable Scheduler - Backend Dependencies
# Compatible with Python 3.11, 3.12, and 3.13

# Core Django packages
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
django-cors-headers>=4.0.0,<5.0.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.0,<6.0.0
django-allauth>=0.57.0,<1.0.0

# Database
dj-database-url>=2.1.0,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0

# Environment Configuration
python-decouple>=3.8,<4.0.0
django-environ>=0.11.0,<1.0.0

# Scheduling and Optimization
ortools>=9.8.0,<10.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0

# File Processing and Export
openpyxl>=3.1.0,<4.0.0
xlsxwriter>=3.1.0,<4.0.0
reportlab>=4.0.0,<5.0.0
Pillow>=10.0.0,<11.0.0
python-dateutil>=2.8.0,<3.0.0

# Background Tasks (Optional)
celery>=5.3.0,<6.0.0
redis>=5.0.0,<6.0.0

# Utilities
python-slugify>=8.0.0,<9.0.0
django-extensions>=3.2.0,<4.0.0
requests>=2.31.0,<3.0.0

# Production
gunicorn>=21.2.0,<22.0.0
whitenoise>=6.6.0,<7.0.0

# Development and Testing
django-debug-toolbar>=4.2.0,<5.0.0
pytest>=7.4.0,<8.0.0
pytest-django>=4.5.0,<5.0.0
factory-boy>=3.3.0,<4.0.0
